# vmmc backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/me/profile": {
      "get": {
        "operationId": "MeController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "MeController_updateProfile",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated key so a retried request never creates a duplicate record.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/compliance-summary": {
      "get": {
        "operationId": "MeController_getComplianceSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/documents": {
      "get": {
        "operationId": "MeController_getMyDocuments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/notifications": {
      "get": {
        "operationId": "MeController_getMyNotifications",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/signature": {
      "get": {
        "operationId": "MeController_getSignature",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "MeController_uploadSignature",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadSignatureDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "delete": {
        "operationId": "MeController_deleteSignature",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/security/pin": {
      "patch": {
        "operationId": "MeController_updatePin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePinDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/security/password": {
      "patch": {
        "operationId": "MeController_updatePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/privacy-settings": {
      "get": {
        "operationId": "MeController_getPrivacySettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "MeController_updatePrivacySettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrivacySettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/performance-stats": {
      "get": {
        "operationId": "MeController_getPerformanceStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/activity-logs": {
      "get": {
        "operationId": "MeController_getActivityLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/login-history": {
      "get": {
        "operationId": "MeController_getLoginHistory",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/devices": {
      "get": {
        "operationId": "MeController_listDevices",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "MeController_registerDevice",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated key so a retried request never creates a duplicate record.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/devices/{id}": {
      "delete": {
        "operationId": "MeController_removeDevice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/documents": {
      "post": {
        "operationId": "DocumentsController_upload",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated key so a retried request never creates a duplicate record.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "examDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "cxrResult": {
                    "type": "string",
                    "enum": [
                      "CLEARED",
                      "INFILTRATE"
                    ]
                  },
                  "genexpertResult": {
                    "type": "string",
                    "enum": [
                      "NOT_DETECTED",
                      "DETECTED"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/pep-logs/{employeeId}": {
      "get": {
        "operationId": "PepLogsController_listForEmployee",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "pep-logs"
        ]
      },
      "post": {
        "operationId": "PepLogsController_create",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated key so a retried request never creates a duplicate record.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePepLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "pep-logs"
        ]
      }
    },
    "/api/pep-logs/{id}": {
      "patch": {
        "operationId": "PepLogsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePepLogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "pep-logs"
        ]
      }
    },
...
    "/api/reports/delinquency": {
      "get": {
        "operationId": "ReportsController_getDelinquency",
        "parameters": [
          {
            "name": "department",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "reports"
        ]
      }
    },
    "/api/reports/clearance-summary": {
      "get": {
        "operationId": "ReportsController_getClearanceSummary",
        "parameters": [
          {
            "name": "department",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "reports"
        ]
      }
    },
    "/api/reports/biological-matrix": {
      "get": {
        "operationId": "ReportsController_getBiologicalMatrix",
        "parameters": [
          {
            "name": "department",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "reports"
        ]
      }
    }
  },
  "info": {
    "title": "VMMC TB DOTS API",
    "description": "ITIL 4 Monitoring & Event Management backend for the VMMC TB DOTS health management system. Write endpoints accept an `Idempotency-Key` header so a client retry after a timeout never duplicates a record.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        }
      },
      "UploadSignatureDto": {
        "type": "object",
        "properties": {
          "imageDataUrl": {
            "type": "string",
            "description": "Base64 PNG data URL, e.g. from canvas.toDataURL(\"image/png\")."
          }
        },
        "required": [
          "imageDataUrl"
        ]
      },
      "UpdatePinDto": {
        "type": "object",
        "properties": {
          "currentPin": {
            "type": "string",
            "description": "Required unless no PIN has ever been set on this account."
          },
          "newPin": {
            "type": "string"
          }
        },
        "required": [
          "newPin"
        ]
      },
      "UpdatePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "UpdatePrivacySettingsDto": {
        "type": "object",
        "properties": {
          "encryptRecords": {
            "type": "boolean"
          },
          "shareAnalytics": {
            "type": "boolean"
          },
          "telemetryLogging": {
            "type": "boolean"
          }
        }
      },
      "CreateDeviceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "iPhone 14 Pro"
          },
          "deviceType": {
            "type": "string",
            "example": "Mobile"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "CreatePepLogDto": {
        "type": "object",
        "properties": {
          "exposureDate": {
            "type": "string",
            "example": "2026-07-01"
          },
          "exposureType": {
            "type": "string",
            "example": "Needlestick injury"
          },
          "prophylaxisStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "IN_PROGRESS",
              "COMPLETED",
              "DECLINED"
            ],
            "default": "PENDING"
          },
          "followUpDate": {
            "type": "string",
            "example": "2026-07-14"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "exposureDate",
          "exposureType"
        ]
      },
      "UpdatePepLogDto": {
        "type": "object",
        "properties": {
          "prophylaxisStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "IN_PROGRESS",
              "COMPLETED",
              "DECLINED"
            ]
          },
          "followUpDate": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
    }
  }
}